challenge_alkemy/requirements.txt
pandas==1.4.2
python-decouple==3.6
requests==2.27.1
SQLAlchemy==1.4.32
SQLAlchemy_Utils==0.37.8
psycopg2==2.8.6
